refactor(api): use directly test-utils/wait.js in tests
