+++ api/tests/llm/integration/domain/usecases/prompt-chat_test.js
@@ -12,15 +12,8 @@ import { promptChat } from '../../../../../src/llm/domain/usecases/prompt-chat.j
 import { chatRepository, promptRepository } from '../../../../../src/llm/infrastructure/repositories/index.js';
 import { LLMResponseHandler } from '../../../../../src/llm/infrastructure/streaming/llm-response-handler.js';
 import { redisMutex } from '../../../../../src/shared/infrastructure/mutex/RedisMutex.js';
-import {
-  catchErr,
-  databaseBuilder,
-  expect,
-  knex,
-  nock,
-  sinon,
-  waitForStreamFinalizationToBeDone,
-} from '../../../../test-helper.js';
+import { catchErr, databaseBuilder, expect, knex, nock, sinon } from '../../../../test-helper.js';
+import { waitForStreamFinalizationToBeDone } from '../../../../tooling/test-utils/wait.js';
 
 describe('LLM | Integration | Domain | UseCases | prompt-chat', function () {
   let dependencies, chatId, llmResponseHandler, resultChunks;
