+++ api/tests/shared/integration/infrastructure/mutex/RedisMutex_test.js
@@ -1,5 +1,6 @@
 import { redisMutex } from '../../../../../src/shared/infrastructure/mutex/RedisMutex.js';
-import { expect, wait } from '../../../../test-helper.js';
+import { expect } from '../../../../test-helper.js';
+import { wait } from '../../../../tooling/test-utils/wait.js';
 
 describe('Shared | Integration | Infrastructure | Mutex | RedisMutex', function () {
   describe('#lock', function () {
